requirements.txt
python-dotenv==1.0.0
openai==1.12.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0 